backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.2
groq==0.4.2
chromadb==0.4.22
sentence-transformers==2.3.1
python-dotenv==1.0.0

Pillow==10.1.0
